+++ Runtime/UnityReactive/ReactiveList/IReadOnlyReactiveList.cs
@@ -3,9 +3,6 @@ using System.Collections.Generic;
 
 namespace Jeomseon.UnityReactive
 {
-    public delegate void ElementChangedHandler<in T>(int index, T previous, T current);
-    public delegate void AddOrRemoveHandler<in T>(int[] indices, T[] items);
-
     /// <summary>
     /// .. 내부 값을 추가/제거/변경이 불가능하고 리스너 추가만 가능한 읽기전용 인터페이스입니다
     /// </summary>