+++ Runtime/UnityReactive/ValueProcessor/IValueProcessor.cs
@@ -1,0 +1,7 @@
+namespace Jeomseon.UnityReactive
+{
+    public interface IValueProcessor
+    {
+        public T Process<T>(T value);
+    }
+}